move pluggable_layer back to custom op
